Fix property struct alignment in OCL, and allow separate log level for OpenCL kernels, set by OCPI_OPENCL_LOG_LEVEL.
adapt/os/ocpios/interfaces/include/OcpiOsDebug.h: change datatype of log level to uint8_t
adapt/os/ocpios/linux/src/OcpiOsDebug.cxx: ditto
core/application/src/ocpirun_main.cxx: ditto
core/hdl/src/ocpihdl_main.cxx: ditto
core/ocl/include/OCL_Worker.h: add a separate "kernelLogLevel" to "self".
core/ocl/include/OclContainer.h: set kernel log level from OCPI_OPENCL_LOG_LEVEL
core/ocl/src/OclContainer.cxx: new kernel log level, and property alignment bug fix

+++ adapt/os/ocpios/interfaces/include/OcpiOsDebug.h
@@ -51,6 +51,7 @@
  *                  Initial version.
  */
 
+#include <stdint.h>
 #include <iostream>
 
 #define OCPI_LOG_DEBUG_MIN 10
@@ -108,8 +109,8 @@ namespace OCPI {
     void debugBreak ()
       throw ();
 
-    void logSetLevel(unsigned n);
-    unsigned logGetLevel();
+    void logSetLevel(uint8_t n);
+    uint8_t logGetLevel();
     void logPrint(unsigned n, const char *fmt, ...) throw() __attribute__((format(printf, 2, 3)));
   }
 }
